Move `BuildIsolation` to `uv-types`
