+++ crates/uv-build/src/lib.rs
@@ -28,12 +28,10 @@ use tracing::{debug, info_span, instrument, Instrument};
 use distribution_types::Resolution;
 use pep440_rs::Version;
 use pep508_rs::{PackageName, Requirement};
+use uv_config::{BuildKind, ConfigSettings, SetupPyStrategy};
 use uv_fs::{PythonExt, Simplified};
 use uv_interpreter::{Interpreter, PythonEnvironment};
-use uv_types::{
-    BuildContext, SourceBuildTrait,
-};
-use uv_config::{BuildIsolation, BuildKind, ConfigSettings, SetupPyStrategy};
+use uv_types::{BuildContext, BuildIsolation, SourceBuildTrait};
 
 /// e.g. `pygraphviz/graphviz_wrap.c:3020:10: fatal error: graphviz/cgraph.h: No such file or directory`
 static MISSING_HEADER_RE: Lazy<Regex> = Lazy::new(|| {
