+++ crates/uv-config/Cargo.toml
@@ -16,7 +16,6 @@ workspace = true
 distribution-types = { workspace = true }
 pep508_rs = { workspace = true }
 uv-cache = { workspace = true }
-uv-interpreter = { workspace = true }
 uv-normalize = { workspace = true }
 
 anyhow = { workspace = true }
