+++ crates/uv-dev/src/resolve_cli.rs
@@ -13,12 +13,12 @@ use distribution_types::{FlatIndexLocation, IndexLocations, IndexUrl, Resolution
 use pep508_rs::Requirement;
 use uv_cache::{Cache, CacheArgs};
 use uv_client::{FlatIndex, FlatIndexClient, RegistryClientBuilder};
+use uv_config::{ConfigSettings, NoBinary, NoBuild, SetupPyStrategy};
 use uv_dispatch::BuildDispatch;
 use uv_installer::SitePackages;
 use uv_interpreter::PythonEnvironment;
 use uv_resolver::{InMemoryIndex, Manifest, Options, Resolver};
-use uv_config::{BuildIsolation, ConfigSettings, NoBinary, NoBuild, SetupPyStrategy};
-use uv_types::InFlight;
+use uv_types::{BuildIsolation, InFlight};
 
 #[derive(ValueEnum, Default, Clone)]
 pub(crate) enum ResolveCliFormat {
