+++ crates/uv-dev/src/resolve_many.rs
@@ -15,12 +15,12 @@ use pep440_rs::{Version, VersionSpecifier, VersionSpecifiers};
 use pep508_rs::{Requirement, VersionOrUrl};
 use uv_cache::{Cache, CacheArgs};
 use uv_client::{FlatIndex, OwnedArchive, RegistryClient, RegistryClientBuilder};
+use uv_config::{ConfigSettings, NoBinary, NoBuild, SetupPyStrategy};
 use uv_dispatch::BuildDispatch;
 use uv_interpreter::PythonEnvironment;
 use uv_normalize::PackageName;
 use uv_resolver::InMemoryIndex;
-use uv_types::{BuildContext, InFlight};
-use uv_config::{NoBinary, BuildIsolation, ConfigSettings, NoBuild, SetupPyStrategy};
+use uv_types::{BuildContext, BuildIsolation, InFlight};
 
 #[derive(Parser)]
 pub(crate) struct ResolveManyArgs {
