+++ crates/uv-dispatch/src/lib.rs
@@ -17,15 +17,11 @@ use pep508_rs::Requirement;
 use uv_build::{SourceBuild, SourceBuildContext};
 use uv_cache::Cache;
 use uv_client::{FlatIndex, RegistryClient};
+use uv_config::{BuildKind, ConfigSettings, NoBinary, NoBuild, Reinstall, SetupPyStrategy};
 use uv_installer::{Downloader, Installer, Plan, Planner, SitePackages};
 use uv_interpreter::{Interpreter, PythonEnvironment};
 use uv_resolver::{InMemoryIndex, Manifest, Options, Resolver};
-use uv_types::{
-    BuildContext, EmptyInstalledPackages, InFlight,
-};
-use uv_config::{BuildIsolation, BuildKind, ConfigSettings,
-    NoBinary, NoBuild, Reinstall, SetupPyStrategy,
-};
+use uv_types::{BuildContext, BuildIsolation, EmptyInstalledPackages, InFlight};
 
 /// The main implementation of [`BuildContext`], used by the CLI, see [`BuildContext`]
 /// documentation.
