+++ crates/uv-types/src/lib.rs
@@ -1,6 +1,8 @@
 //! Fundamental types shared across `uv` crates.
+pub use builds::*;
 pub use downloads::*;
 pub use traits::*;
 
+mod builds;
 mod downloads;
 mod traits;
