+++ crates/uv/src/commands/pip_sync.rs
@@ -19,9 +19,7 @@ use uv_client::{
     BaseClientBuilder, Connectivity, FlatIndex, FlatIndexClient, RegistryClient,
     RegistryClientBuilder,
 };
-use uv_config::{
-    BuildIsolation, ConfigSettings, IndexStrategy, NoBinary, NoBuild, Reinstall, SetupPyStrategy,
-};
+use uv_config::{ConfigSettings, IndexStrategy, NoBinary, NoBuild, Reinstall, SetupPyStrategy};
 use uv_dispatch::BuildDispatch;
 use uv_fs::Simplified;
 use uv_installer::{is_dynamic, Downloader, Plan, Planner, ResolvedEditable, SitePackages};
@@ -31,7 +29,7 @@ use uv_requirements::{
     SourceTreeResolver,
 };
 use uv_resolver::{DependencyMode, InMemoryIndex, Manifest, OptionsBuilder, Resolver};
-use uv_types::{EmptyInstalledPackages, InFlight};
+use uv_types::{BuildIsolation, EmptyInstalledPackages, InFlight};
 use uv_warnings::warn_user;
 
 use crate::commands::reporters::{DownloadReporter, InstallReporter, ResolverReporter};
